continue home working

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -8,10 +8,12 @@ import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
 
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
 @Repository
 public class InMemoryMealRepositoryImpl implements MealRepository {
@@ -21,38 +23,42 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
 
     {
         MealsUtil.MEALS.forEach(m -> save(m, 2));
-
     }
 
     @Override
-    public Meal save(Meal meal, int idUser) {
+    public Meal save(Meal meal, int userId) {
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
-            meal.setOwnerUserId(idUser);
+            meal.setOwnerUserId(userId);
         }
         repository.put(meal.getId(), meal);
+        log.debug("is puting {}", meal);
         return meal;
     }
 
     @Override
-    public void delete(int id) {
-        repository.remove(id);
+    public void delete(int id, int userId) {
+        log.info("prepare to delete meal:{} by userId{}", repository.get(id), userId);
+        if (repository.get(id).getOwnerUserId() == userId) repository.remove(id);
     }
 
     @Override
-    public Meal get(int id) {
-        return repository.get(id);
+    public Meal get(int id, int userId) {
+        log.info("get meal[id={}] from user[id={}]", id, userId);
+        return (repository.get(id).getOwnerUserId() == userId) ? repository.get(id) : null;
     }
 
     @Override
-    public List<Meal> getAll() {
-        log.info("get all");
-        List<Meal> meals = new ArrayList<>();
-        meals.addAll(repository.values());
-        log.debug("success conver to list<Meal>");
-        meals.sort((meal, t1) -> -1 * meal.getDateTime().compareTo(t1.getDateTime()));
-        log.debug("all:{}", meals);
-        return meals;
+    public List<Meal> getAll(int idUser) {
+        log.info("get all by user[id={}]", idUser);
+        List<Meal> mealList = new ArrayList<>();
+        mealList.addAll(repository.values());
+        log.debug("is initialization completed. mealList:{}", mealList);
+        List<Meal> result = mealList.stream()
+                .filter(meal -> (meal.getOwnerUserId() != null && meal.getOwnerUserId() == idUser))
+                .sorted((x, y) -> -1 * (x.getDateTime().compareTo(y.getDateTime())))
+                .collect(Collectors.toList());
+        log.debug("return result:{}", result);
+        return result;
     }
 }
-
